neo4j.CachedDb -> EventBuffer
