+++ dagreader/go.mod
@@ -10,6 +10,7 @@ require (
 	github.com/neo4j/neo4j-go-driver v1.8.3
 	github.com/paulbellamy/ratecounter v0.2.0
 	github.com/stretchr/testify v1.7.0
+	github.com/syndtr/goleveldb v1.0.1-0.20210305035536-64b5b1c73954
 	github.com/urfave/cli v1.22.1
 	gopkg.in/urfave/cli.v1 v1.22.1 // indirect
 )
